Transaction Number

+++ team8finalproject/Utilities/GenerateTransactionNumber.cs
@@ -9,8 +9,8 @@ namespace team8finalproject.Utilities
     {
         public static Int64 GetNextTransactionNumber(AppDbContext _context)
         {
-            Int32 intMaxTransactionNumber; //the current maximum number
-            Int32 intNextTransactionNumber; //the trans number
+            Int64 intMaxTransactionNumber; //the current maximum number
+            Int64 intNextTransactionNumber; //the trans number
 
             if (_context.Transactions.Count() == 0) //there are no transactions in the database yet
             {
